webrtc_sctp::stream

Struct Stream

pub struct Stream { /* private fields */ }

Stream represents an SCTP stream

Implementations

impl Stream

pub fn stream_identifier(&self) -> u16

stream_identifier returns the Stream identifier associated to the stream.

pub fn set_default_payload_type( &self, default_payload_type: PayloadProtocolIdentifier, )

set_default_payload_type sets the default payload type used by write.

pub fn set_reliability_params( &self, unordered: bool, rel_type: ReliabilityType, rel_val: u32, )

set_reliability_params sets reliability parameters for this stream.

pub async fn read(&self, p: &mut [u8]) -> Result<usize, Error>

Reads a packet of len(p) bytes, dropping the Payload Protocol Identifier.

Returns Error::ErrShortBuffer if p is too short. Returns 0 if the reading half of this stream is shutdown or it (the stream) was reset.

pub async fn read_sctp( &self, p: &mut [u8], ) -> Result<(usize, PayloadProtocolIdentifier), Error>

Reads a packet of len(p) bytes and returns the associated Payload Protocol Identifier.

Returns Error::ErrShortBuffer if p is too short. Returns (0, PayloadProtocolIdentifier::Unknown) if the reading half of this stream is shutdown or it (the stream) was reset.

pub async fn write(&self, p: &Bytes) -> Result<usize, Error>

Writes p to the DTLS connection with the default Payload Protocol Identifier.

Returns an error if the write half of this stream is shutdown or p is too large.

pub async fn write_sctp( &self, p: &Bytes, ppi: PayloadProtocolIdentifier, ) -> Result<usize, Error>

Writes p to the DTLS connection with the given Payload Protocol Identifier.

Returns an error if the write half of this stream is shutdown or p is too large.

pub async fn close(&self) -> Result<(), Error>

👎Deprecated

Closes both read and write halves of this stream.

Use Stream::shutdown instead.

pub async fn shutdown(&self, how: Shutdown) -> Result<(), Error>

Shuts down the read, write, or both halves of this stream.

This function will cause all pending and future I/O on the specified portions to return immediately with an appropriate value (see the documentation of Shutdown).

Resets the stream when both halves of this stream are shutdown.

pub fn buffered_amount(&self) -> usize

buffered_amount returns the number of bytes of data currently queued to be sent over this stream.

pub fn buffered_amount_low_threshold(&self) -> usize

buffered_amount_low_threshold returns the number of bytes of buffered outgoing data that is considered “low.” Defaults to 0.

pub fn set_buffered_amount_low_threshold(&self, th: usize)

set_buffered_amount_low_threshold is used to update the threshold. See buffered_amount_low_threshold().

pub fn on_buffered_amount_low(&self, f: OnBufferedAmountLowFn)

on_buffered_amount_low sets the callback handler which would be called when the number of bytes of outgoing data buffered is lower than the threshold.